updates heroicon helper
The gem "heroicon" hasn't been updated in 18mo, and the heroicon pack has dramatically changed since then.

Another gem "rails-heroicon" seems to share a lot of the code but is actually bundled with a railtie and has carried updates as recently as 2 weeks ago. The interface is identical from the views, but it comes with all the names for icons which match the current heroicon website.

I've removed the heroicon helper and initializer. Neither are required by rails-heroicon.